refactor: Rename color/test formatter to color/test highlighter
