+++ test/helpers/test-highlighter.js
